fix(host-service): route getDiffBulk's git spawns off the main loop
CI's no-main-loop-blocking ratchet caught it: getDiffBulk constructed a
SimpleGit client per file directly in the tRPC resolver, adding 2 new
on-loop createUserSimpleGit call sites to git.ts (4 > the allowed 2).
Each spawn's stdout drain runs on host-service's event loop, which is
also serving every other tRPC response, status poll, and watcher
callback in the process — exactly the head-of-line blocking the ratchet
exists to prevent.

Moves resolveDiffCategoryRefs/loadFileDiffContent into git-helpers.ts
(shared, ratchet-neutral — they take a `git` client, they don't
construct one) and adds a gitDiffBulkTask worker task that resolves
refs and loads every file's diff with bounded concurrency inside the
worker pool, mirroring the existing getStatus/getCommitFiles pattern.
getDiffBulk's resolver now just resolves gitEnv on-loop and hands off
to the pool.

Verified: no-main-loop-blocking.test.ts passes, full host-service suite
(1157 pass) and the 300-file get-diff-bulk.test.ts both green.
